Fix handling of unknown options and missing args
In the error message when unknown option is encountered, it only show
the short options (with single dash '-') and not the long options
(the one with two dash '--').

Having a confusing return value of getopt_long() function when an option
that requires an argument have no argument provided by user. See this
page: https://linux.die.net/man/3/getopt, in the 'Return value'.

To address the first issue, I use the variable 'optopt' to identify if
the unknown option is short or a long. In the second issue, I put a
colon (':') as the first character in the 'optstring' of getopt_long() to
uniquely identify the error if it is a missing argument of an option or
an unknown option.

The question mark character '?' is returned by getopt_long() function when
an option not in the 'opstring' is found and the colon ':' character is
returned when an option that requires argument does not have.